Refactor submit hooks
